[chore][pkg/stanza] Cleanup regex parser operator files
