+++ pkg/stanza/operator/parser/regex/parser_test.go
